+++ packages/contracts/contracts/libs/WalletUtils.sol
@@ -6,21 +6,34 @@ import {Utils} from "../libs/Utils.sol";
 
 // Helpers for Wallet.sol
 library WalletUtils {
-    function extractJoinSplitProofsAndPisFromBundle(
-        IWallet.Bundle calldata bundle
+    function computeOperationDigests(
+        IWallet.Operation[] calldata ops
+    ) internal pure returns (uint256[] memory) {
+        uint256 numOps = ops.length;
+        uint256[] memory opDigests = new uint256[](numOps);
+        for (uint256 i = 0; i < numOps; i++) {
+            opDigests[i] = computeOperationDigest(ops[i]);
+        }
+
+        return opDigests;
+    }
+
+    function extractJoinSplitProofsAndPis(
+        IWallet.Operation[] calldata ops,
+        uint256[] memory digests
     )
         internal
         pure
         returns (Groth16.Proof[] memory proofs, uint256[][] memory allPis)
     {
-        uint256 numOps = bundle.operations.length;
+        uint256 numOps = ops.length;
 
         // compute number of joinsplits in the bundle
         uint256 numJoinSplits = 0;
         for (uint256 i = 0; i < numOps; i++) {
-            IWallet.Operation calldata op = bundle.operations[i];
-            numJoinSplits += op.joinSplitTxs.length;
+            numJoinSplits += ops[i].joinSplitTxs.length;
         }
+
         proofs = new Groth16.Proof[](numJoinSplits);
         allPis = new uint256[][](numJoinSplits);
 
@@ -29,9 +42,7 @@ library WalletUtils {
 
         // Batch verify all the joinsplit proofs
         for (uint256 i = 0; i < numOps; i++) {
-            IWallet.Operation calldata op = bundle.operations[i];
-            uint256 operationDigest = uint256(_hashOperation(op)) %
-                Utils.SNARK_SCALAR_FIELD;
+            IWallet.Operation memory op = ops[i];
             for (uint256 j = 0; j < op.joinSplitTxs.length; j++) {
                 proofs[index] = Utils.proof8ToStruct(op.joinSplitTxs[j].proof);
                 allPis[index] = new uint256[](9);
@@ -41,7 +52,7 @@ library WalletUtils {
                 allPis[index][3] = op.joinSplitTxs[j].publicSpend;
                 allPis[index][4] = op.joinSplitTxs[j].nullifierA;
                 allPis[index][5] = op.joinSplitTxs[j].nullifierB;
-                allPis[index][6] = operationDigest;
+                allPis[index][6] = digests[i];
                 allPis[index][7] = uint256(uint160(op.joinSplitTxs[j].asset));
                 allPis[index][8] = op.joinSplitTxs[j].id;
                 index++;
@@ -53,12 +64,12 @@ library WalletUtils {
 
     // TODO: do we need a domain in the payload?
     // TODO: turn encodedFunctions and contractAddresses into their own arrays, so we don't have to call abi.encodePacked for each one
-    function _hashOperation(
+    function computeOperationDigest(
         IWallet.Operation calldata op
-    ) private pure returns (bytes32) {
+    ) internal pure returns (uint256) {
         bytes memory payload;
 
-        IWallet.Action calldata action;
+        IWallet.Action memory action;
         for (uint256 i = 0; i < op.actions.length; i++) {
             action = op.actions[i];
             payload = abi.encodePacked(
@@ -95,7 +106,7 @@ library WalletUtils {
             op.gasLimit
         );
 
-        return keccak256(payload);
+        return uint256(keccak256(payload)) % Utils.SNARK_SCALAR_FIELD;
     }
 
     function _hashJoinSplit(