+++ packages/subtree-updater/src/rapidsnarkProver.ts
@@ -1,4 +1,9 @@
-import { SubtreeUpdateProver, SubtreeUpdateInputs, SubtreeUpdateProofWithPublicSignals, toJSON } from "@nocturne-xyz/sdk";
+import {
+  SubtreeUpdateProver,
+  SubtreeUpdateInputs,
+  SubtreeUpdateProofWithPublicSignals,
+  toJSON,
+} from "@nocturne-xyz/sdk";
 
 //@ts-ignore
 import * as snarkjs from "snarkjs";
@@ -6,13 +11,19 @@ import * as fs from "fs";
 import { spawn } from "child_process";
 
 export class RapidsnarkSubtreeUpdateProver implements SubtreeUpdateProver {
-  rapidsnarkExecutablePath: string; 
+  rapidsnarkExecutablePath: string;
   witnessGeneratorExecutablePath: string;
   zkeyPath: string;
   tmpDir: string;
   vkey: any;
 
-  constructor(rapidsnarkExecutablePath: string, witnessGeneratorExecutablePath: string, zkeyPath: string, vkeyPath: string, tmpDir: string = __dirname) {
+  constructor(
+    rapidsnarkExecutablePath: string,
+    witnessGeneratorExecutablePath: string,
+    zkeyPath: string,
+    vkeyPath: string,
+    tmpDir: string = __dirname
+  ) {
     this.rapidsnarkExecutablePath = rapidsnarkExecutablePath;
     this.witnessGeneratorExecutablePath = witnessGeneratorExecutablePath;
     this.zkeyPath = zkeyPath;
@@ -20,16 +31,22 @@ export class RapidsnarkSubtreeUpdateProver implements SubtreeUpdateProver {
     this.vkey = JSON.parse(fs.readFileSync(vkeyPath).toString());
   }
 
-  async proveSubtreeUpdate(inputs: SubtreeUpdateInputs): Promise<SubtreeUpdateProofWithPublicSignals> {
+  async proveSubtreeUpdate(
+    inputs: SubtreeUpdateInputs
+  ): Promise<SubtreeUpdateProofWithPublicSignals> {
     const inputJsonPath = `${this.tmpDir}/_input.json`;
     const witnessPath = `${this.tmpDir}/_witness.wtns`;
     const proofJsonPath = `${this.tmpDir}/_proof.json`;
     const publicSignalsPath = `${this.tmpDir}/_public.json`;
 
     await fs.promises.writeFile(inputJsonPath, toJSON(inputs));
-    await runCommand(`${this.witnessGeneratorExecutablePath} ${inputJsonPath} ${witnessPath}`);
-    await runCommand(`${this.rapidsnarkExecutablePath} ${this.zkeyPath} ${witnessPath} ${proofJsonPath} ${publicSignalsPath}`);
-    
+    await runCommand(
+      `${this.witnessGeneratorExecutablePath} ${inputJsonPath} ${witnessPath}`
+    );
+    await runCommand(
+      `${this.rapidsnarkExecutablePath} ${this.zkeyPath} ${witnessPath} ${proofJsonPath} ${publicSignalsPath}`
+    );
+
     const [proofStr, publicSignalsStr] = await Promise.all([
       fs.promises.readFile(proofJsonPath, "utf-8"),
       fs.promises.readFile(publicSignalsPath, "utf-8"),
@@ -37,32 +54,35 @@ export class RapidsnarkSubtreeUpdateProver implements SubtreeUpdateProver {
 
     const proof = JSON.parse(proofStr);
     const publicSignals = JSON.parse(publicSignalsStr);
-    
+
     return {
       proof,
       publicSignals,
     };
   }
 
-  async verifySubtreeUpdate({ proof, publicSignals }: SubtreeUpdateProofWithPublicSignals): Promise<boolean> {
+  async verifySubtreeUpdate({
+    proof,
+    publicSignals,
+  }: SubtreeUpdateProofWithPublicSignals): Promise<boolean> {
     return await snarkjs.groth16.verify(this.vkey, publicSignals, proof);
   }
 }
 
 async function runCommand(cmd: string): Promise<[string, string]> {
   return new Promise((resolve, reject) => {
-    let stdout = '';
-    let stderr = '';
-    const child = spawn('sh', ['-c', cmd]);
-    child.stdout.on('data', data => {
+    let stdout = "";
+    let stderr = "";
+    const child = spawn("sh", ["-c", cmd]);
+    child.stdout.on("data", (data) => {
       const output = data.toString();
       stdout += output;
     });
-    child.stderr.on('data', data => {
+    child.stderr.on("data", (data) => {
       const output = data.toString();
       stderr += output;
     });
-    child.on('error', reject);
-    child.on('exit', () => resolve([stdout, stderr]));
+    child.on("error", reject);
+    child.on("exit", () => resolve([stdout, stderr]));
   });
 }