+++ packages/subtree-updater/src/submitter.ts
@@ -2,7 +2,11 @@ import { Wallet } from "@nocturne-xyz/contracts";
 import { BaseProof, packToSolidityProof } from "@nocturne-xyz/sdk";
 
 export interface SubtreeUpdateSubmitter {
-  submitProof (proof: BaseProof, newRoot: bigint, subtreeIndex: number): Promise<void>;
+  submitProof(
+    proof: BaseProof,
+    newRoot: bigint,
+    subtreeIndex: number
+  ): Promise<void>;
   dropDB(): Promise<void>;
 }
 
@@ -18,10 +22,12 @@ export class SyncSubtreeSubmitter implements SubtreeUpdateSubmitter {
   async submitProof(proof: BaseProof, newRoot: bigint): Promise<void> {
     const solidityProof = packToSolidityProof(proof);
     try {
-      const tx = await this.walletContract.applySubtreeUpdate(newRoot, solidityProof); 
+      const tx = await this.walletContract.applySubtreeUpdate(
+        newRoot,
+        solidityProof
+      );
       await tx.wait();
     } catch (err: any) {
-
       // ignore errors that are due to duplicate submissions
       // this can happen if there are multiple instances of subtree updaters running
       if (!err.toString().includes("newRoot already a past root")) {